+++ src/embed_tests/Codecs.cs
@@ -356,13 +356,38 @@ DateTimeDecoder.Setup()
         }
 
         [Test]
+        public void FloatDerivedDecoded()
+        {
+            using var scope = Py.CreateScope();
+            scope.Exec(@"class FloatDerived(float): pass");
+            using var floatDerived = scope.Eval("FloatDerived");
+            var decoder = new DecoderReturningPredefinedValue<object>(floatDerived, 42);
+            PyObjectConversions.RegisterDecoder(decoder);
+            using var result = scope.Eval("FloatDerived()");
+            object decoded = result.As<object>();
+            Assert.AreEqual(42, decoded);
+        }
+
+        [Test]
         public void ExceptionDecodedNoInstance()
         {
-            PyObjectConversions.RegisterDecoder(new InstancelessExceptionDecoder());
-            using var scope = Py.CreateScope();
-            var error = Assert.Throws<ValueErrorWrapper>(() => PythonEngine.Exec(
-                $"[].__iter__().__next__()"));
-            Assert.AreEqual(TestExceptionMessage, error.Message);
+            if (Runtime.PyVersion < new Version(3, 12))
+            {
+                PyObjectConversions.RegisterDecoder(new InstancelessExceptionDecoder());
+                using var scope = Py.CreateScope();
+
+                var error = Assert.Throws<ValueErrorWrapper>(() =>
+                    PythonEngine.Exec($"[].__iter__().__next__()")
+                );
+                Assert.AreEqual(TestExceptionMessage, error.Message);
+            }
+            else
+            {
+                Assert.Ignore(
+                    "This test does not work for Python 3.12, see " +
+                    "https://github.com/python/cpython/issues/101578"
+                );
+            }
         }
 
         public static void AcceptsDateTime(DateTime v) {}