+++ src/runtime/Converter.cs
@@ -361,6 +361,29 @@ namespace Python.Runtime
             // conversions (Python string -> managed string).
             if (obType == objectType)
             {
+                if (Runtime.PyString_CheckExact(value))
+                {
+                    return ToPrimitive(value, stringType, out result, setError);
+                }
+
+                if (Runtime.PyBool_CheckExact(value))
+                {
+                    return ToPrimitive(value, boolType, out result, setError);
+                }
+
+                if (Runtime.PyFloat_CheckExact(value))
+                {
+                    return ToPrimitive(value, doubleType, out result, setError);
+                }
+
+                // give custom codecs a chance to take over conversion
+                // of ints, sequences, and types derived from primitives
+                BorrowedReference pyType = Runtime.PyObject_TYPE(value);
+                if (PyObjectConversions.TryDecode(value, pyType, obType, out result))
+                {
+                    return true;
+                }
+
                 if (Runtime.PyString_Check(value))
                 {
                     return ToPrimitive(value, stringType, out result, setError);
@@ -374,13 +397,6 @@ namespace Python.Runtime
                 if (Runtime.PyFloat_Check(value))
                 {
                     return ToPrimitive(value, doubleType, out result, setError);
-                }
-
-                // give custom codecs a chance to take over conversion of ints and sequences
-                BorrowedReference pyType = Runtime.PyObject_TYPE(value);
-                if (PyObjectConversions.TryDecode(value, pyType, obType, out result))
-                {
-                    return true;
                 }
 
                 if (Runtime.PyInt_Check(value))
@@ -670,10 +686,8 @@ namespace Python.Runtime
                         {
                             if (Runtime.PyUnicode_GetLength(value) == 1)
                             {
-                                IntPtr unicodePtr = Runtime.PyUnicode_AsUnicode(value);
-                                Char[] buff = new Char[1];
-                                Marshal.Copy(unicodePtr, buff, 0, 1);
-                                result = buff[0];
+                                int chr = Runtime.PyUnicode_ReadChar(value, 0);
+                                result = (Char)chr;
                                 return true;
                             }
                             goto type_error;