+++ src/runtime/Runtime.cs
@@ -598,23 +598,8 @@ namespace Python.Runtime
         [Obsolete("Use NewReference or PyObject constructor instead")]
         internal static unsafe void XIncref(BorrowedReference op)
         {
-#if !CUSTOM_INCDEC_REF
             Py_IncRef(op);
             return;
-#else
-            var p = (void*)op;
-            if ((void*)0 != p)
-            {
-                if (Is32Bit)
-                {
-                    (*(int*)p)++;
-                }
-                else
-                {
-                    (*(long*)p)++;
-                }
-            }
-#endif
         }
 
         internal static unsafe void XDecref(StolenReference op)
@@ -623,40 +608,9 @@ namespace Python.Runtime
             Debug.Assert(op == null || Refcount(new BorrowedReference(op.Pointer)) > 0);
             Debug.Assert(_isInitialized || Py_IsInitialized() != 0 || _Py_IsFinalizing() != false);
 #endif
-#if !CUSTOM_INCDEC_REF
             if (op == null) return;
             Py_DecRef(op.AnalyzerWorkaround());
             return;
-#else
-            var p = (void*)op;
-            if ((void*)0 != p)
-            {
-                if (Is32Bit)
-                {
-                    --(*(int*)p);
-                }
-                else
-                {
-                    --(*(long*)p);
-                }
-                if ((*(int*)p) == 0)
-                {
-                    // PyObject_HEAD: struct _typeobject *ob_type
-                    void* t = Is32Bit
-                        ? (void*)(*((uint*)p + 1))
-                        : (void*)(*((ulong*)p + 1));
-                    // PyTypeObject: destructor tp_dealloc
-                    void* f = Is32Bit
-                        ? (void*)(*((uint*)t + 6))
-                        : (void*)(*((ulong*)t + 6));
-                    if ((void*)0 == f)
-                    {
-                        return;
-                    }
-                    NativeCall.Void_Call_1(new IntPtr(f), op);
-                }
-            }
-#endif
         }
 
         [Pure]
@@ -671,6 +625,9 @@ namespace Python.Runtime
         }
         [Pure]
         internal static int Refcount32(BorrowedReference op) => checked((int)Refcount(op));
+
+        internal static void TryUsingDll(Action op) =>
+            TryUsingDll(() => { op(); return 0; });
 
         /// <summary>
         /// Call specified function, and handle PythonDLL-related failures.
@@ -982,7 +939,7 @@ namespace Python.Runtime
             Debug.Assert(ob != null);
             var type = PyObject_TYPE(ob);
             int offset = Util.ReadInt32(type, TypeOffset.tp_weaklistoffset);
-            if (offset == 0) return BorrowedReference.Null;
+            if (offset <= 0) return BorrowedReference.Null;
             Debug.Assert(offset > 0);
             return Util.ReadRef(ob, offset);
         }
@@ -1094,8 +1051,13 @@ namespace Python.Runtime
         internal static bool PyInt_Check(BorrowedReference ob)
             => PyObject_TypeCheck(ob, PyLongType);
 
+        internal static bool PyInt_CheckExact(BorrowedReference ob)
+            => PyObject_TypeCheckExact(ob, PyLongType);
+
         internal static bool PyBool_Check(BorrowedReference ob)
             => PyObject_TypeCheck(ob, PyBoolType);
+        internal static bool PyBool_CheckExact(BorrowedReference ob)
+            => PyObject_TypeCheckExact(ob, PyBoolType);
 
         internal static NewReference PyInt_FromInt32(int value) => PyLong_FromLongLong(value);
 
@@ -1141,6 +1103,8 @@ namespace Python.Runtime
 
         internal static bool PyFloat_Check(BorrowedReference ob)
             => PyObject_TypeCheck(ob, PyFloatType);
+        internal static bool PyFloat_CheckExact(BorrowedReference ob)
+            => PyObject_TypeCheckExact(ob, PyFloatType);
 
         /// <summary>
         /// Return value: New reference.
@@ -1282,9 +1246,9 @@ namespace Python.Runtime
         // Python string API
         //====================================================================
         internal static bool PyString_Check(BorrowedReference ob)
-        {
-            return PyObject_TYPE(ob) == PyStringType;
-        }
+            => PyObject_TypeCheck(ob, PyStringType);
+        internal static bool PyString_CheckExact(BorrowedReference ob)
+            => PyObject_TypeCheckExact(ob, PyStringType);
 
         internal static NewReference PyString_FromString(string value)
         {
@@ -1326,8 +1290,9 @@ namespace Python.Runtime
         internal static nint PyUnicode_GetLength(BorrowedReference ob) => Delegates.PyUnicode_GetLength(ob);
 
 
-        internal static IntPtr PyUnicode_AsUnicode(BorrowedReference ob) => Delegates.PyUnicode_AsUnicode(ob);
         internal static NewReference PyUnicode_AsUTF16String(BorrowedReference ob) => Delegates.PyUnicode_AsUTF16String(ob);
+
+        internal static int PyUnicode_ReadChar(BorrowedReference ob, nint index) => Delegates.PyUnicode_ReadChar(ob, index);
 
 
 
@@ -1643,6 +1608,8 @@ namespace Python.Runtime
             return Delegates.PyType_IsSubtype(t1, t2);
         }
 
+        internal static bool PyObject_TypeCheckExact(BorrowedReference ob, BorrowedReference tp)
+            => PyObject_TYPE(ob) == tp;
         internal static bool PyObject_TypeCheck(BorrowedReference ob, BorrowedReference tp)
         {
             BorrowedReference t = PyObject_TYPE(ob);