+++ src/runtime/Types/MetaType.cs
@@ -82,52 +82,69 @@ namespace Python.Runtime
             BorrowedReference bases = Runtime.PyTuple_GetItem(args, 1);
             BorrowedReference dict = Runtime.PyTuple_GetItem(args, 2);
 
-            // We do not support multiple inheritance, so the bases argument
-            // should be a 1-item tuple containing the type we are subtyping.
-            // That type must itself have a managed implementation. We check
-            // that by making sure its metatype is the CLR metatype.
+            // Extract interface types and base class types.
+            var interfaces = new List<Type>();
 
+            // More than one base type case be declared, but an exception will be thrown
+            // if more than one is a class/not an interface.
+            var baseTypes = new List<ClassBase>();
 
-            List<Type> interfaces = new List<Type>();
-            List<ClassBase> baseType = new List<ClassBase>();
-
-            var cnt = Runtime.PyTuple_GetSize(bases);
-
-            for (uint i = 0; i < cnt; i++)
+            var baseClassCount = Runtime.PyTuple_Size(bases);
+            if (baseClassCount == 0)
             {
-                var base_type2 = Runtime.PyTuple_GetItem(bases, (int)i);
-                var cb2 = (ClassBase) GetManagedObject(base_type2);
-                if (cb2 != null)
+                return Exceptions.RaiseTypeError("zero base classes ");
+            }
+
+            for (nint i = 0; i < baseClassCount; i++)
+            {
+                var baseTypeIt = Runtime.PyTuple_GetItem(bases, (int)i);
+
+                if (GetManagedObject(baseTypeIt) is ClassBase classBaseIt)
                 {
-                    if (cb2.type.Valid && cb2.type.Value.IsInterface)
-                        interfaces.Add(cb2.type.Value);
-                    else baseType.Add(cb2);
+                    if (!classBaseIt.type.Valid)
+                    {
+                        return Exceptions.RaiseTypeError("Invalid type used as a super type.");
+                    }
+                    if (classBaseIt.type.Value.IsInterface)
+                    {
+                        interfaces.Add(classBaseIt.type.Value);
+                    }
+                    else
+                    {
+                        baseTypes.Add(classBaseIt);
+                    }
+                }
+                else
+                {
+                    return Exceptions.RaiseTypeError("Non .NET type used as super class for meta type. This is not supported.");
                 }
             }
-
-            if (baseType.Count == 0)
+            // if the base type count is 0, there might still be interfaces to implement.
+            if (baseTypes.Count == 0)
             {
-                baseType.Add(new ClassBase(typeof(object)));
+                baseTypes.Add(new ClassBase(typeof(object)));
             }
 
-
-            if (baseType.Count > 1)
+            // Multiple inheritance is not supported, unless the other types are interfaces
+            if (baseTypes.Count > 1)
             {
-                return Exceptions.RaiseTypeError("cannot use multiple inheritance with managed classes");
+                var types = string.Join(", ", baseTypes.Select(baseType => baseType.type.Value));
+                return Exceptions.RaiseTypeError($"Multiple inheritance with managed classes cannot be used. Types: {types} ");
             }
 
-            /*
-            BorrowedReference mt = Runtime.PyObject_TYPE(baseType);
-
-            if (!(mt == PyCLRMetaType || mt == Runtime.PyTypeType))
+            // check if the list of interfaces contains no duplicates.
+            if (interfaces.Distinct().Count() != interfaces.Count)
             {
-                return Exceptions.RaiseTypeError("invalid metatype");
-            }*/
+                // generate a string containing the problematic types.
+                var duplicateTypes = interfaces.GroupBy(type => type)
+                    .Where(typeGroup => typeGroup.Count() > 1)
+                    .Select(typeGroup => typeGroup.Key);
+                var duplicateTypesString = string.Join(", ", duplicateTypes);
 
-            // Ensure that the reflected type is appropriate for subclassing,
-            // disallowing subclassing of delegates, enums and array types.
+                return Exceptions.RaiseTypeError($"An interface can only be implemented once. Duplicate types: {duplicateTypesString}");
+            }
 
-            var cb = baseType.First();
+            var cb = baseTypes[0];
             try
             {
                 if (!cb.CanSubclass())
@@ -152,11 +169,11 @@ namespace Python.Runtime
             // into python.
             if (null != dict)
             {
-                var btt = baseType.FirstOrDefault().type.ValueOrNull;
+                var btt = baseTypes.FirstOrDefault().type.ValueOrNull;
                 var ctor = btt?.GetConstructors(BindingFlags.Instance | BindingFlags.Public | BindingFlags.NonPublic)
                     .FirstOrDefault(x => x.GetParameters().Any() == false);
                 using var clsDict = new PyDict(dict);
-                
+
                 if (clsDict.HasKey("__assembly__") || clsDict.HasKey("__namespace__")
                                                    || (ctor != null))
                 {
@@ -165,9 +182,11 @@ namespace Python.Runtime
                         clsDict["__namespace__"] =
                             (clsDict["__module__"].ToString()).ToPython();
                     }
-                    return TypeManager.CreateSubType(name, baseType, interfaces, clsDict);
+
+                    return TypeManager.CreateSubType(name, baseTypes[0], interfaces, clsDict);
+
                 }
-                
+
             }
 
             var base_type = Runtime.PyTuple_GetItem(bases, 0);