+++ src/testing/classtest.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections;
 
 namespace Python.Test