+++ tests/test_subclass.py
@@ -12,8 +12,7 @@ from System.Diagnostics import (DebuggerDisplay, DebuggerDisplayAttribute, Debug
 from System.ComponentModel import (Browsable, BrowsableAttribute)
 from System.Threading import (CancellationToken)
 import pytest
-from Python.Test import (IInterfaceTest, SubClassTest, EventArgsTest,
-                         FunctionsTest, GenericVirtualMethodTest, ISimpleInterface, SimpleClass, TestAttribute, TestAttributeAttribute, ISimpleInterface2)
+from Python.Test import (IInterfaceTest, SubClassTest, EventArgsTest, FunctionsTest,IGenericInterface, GenericVirtualMethodTest, ISimpleInterface, SimpleClass, TestAttribute, TestAttributeAttribute, ISimpleInterface2, ISayHello1)
 import Python.Test
 from System.Collections.Generic import List
 
@@ -32,6 +31,17 @@ def interface_test_class_fixture(subnamespace):
             return "/".join([x] * i)
 
     return InterfaceTestClass
+
+
+def interface_generic_class_fixture(subnamespace):
+
+    class GenericInterfaceImpl(IGenericInterface[int]):
+        __namespace__ = "Python.Test." + subnamespace
+
+        def Get(self, x):
+            return x
+
+    return GenericInterfaceImpl
 
 
 def derived_class_fixture(subnamespace):
@@ -489,3 +499,30 @@ def test_more_subclasses2():
     print(x.CallIncrementThing())
 
 
+
+def test_generic_interface():
+    from System import Int32
+    from Python.Test import GenericInterfaceUser, SpecificInterfaceUser
+
+    GenericInterfaceImpl = interface_generic_class_fixture(test_generic_interface.__name__)
+
+    obj = GenericInterfaceImpl()
+    SpecificInterfaceUser(obj, Int32(0))
+    GenericInterfaceUser[Int32](obj, Int32(0))
+
+def test_virtual_generic_method():
+    class OverloadingSubclass(GenericVirtualMethodTest):
+        __namespace__ = "test_virtual_generic_method_cls"
+    class OverloadingSubclass2(OverloadingSubclass):
+        __namespace__ = "test_virtual_generic_method_cls"
+    obj = OverloadingSubclass()
+    assert obj.VirtMethod[int](5) == 5
+    obj = OverloadingSubclass2()
+    assert obj.VirtMethod[int](5) == 5
+
+def test_implement_interface_and_class():
+    class DualSubClass0(ISayHello1, SimpleClass):
+        __namespace__ = "Test"
+        def SayHello(self):
+            return "hello"
+    obj = DualSubClass0()